requirements.txt
aiofiles==0.7.0
arrow==1.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
attrs==21.2.0
Automat==20.2.0
Beaker==1.11.0
binaryornot==0.4.4
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
click==7.1.2
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
constantly==15.1.0
cookiecutter==1.7.2
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cssselect==1.1.0
dbus-python==1.2.16
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-tailwind==2.0.1
fastapi==0.65.1
fildem @ git+https://github.com/gonzaarcr/Fildem@9fd2dbf6b87bf0d5539a7b66eb6d6052534fa1dc
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
img2pdf==0.4.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
parsel==1.6.0
pikepdf==2.11.1
Pillow==8.2.0
poyo==0.5.0
priority==1.3.0
Protego==0.1.16
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298147424/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.8.2
PyDispatcher==2.0.5
PyGObject==3.40.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.1
python-slugify==5.0.2
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
qrcode==6.1
queuelib==1.6.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
Scrapy==2.5.0
service-identity==18.1.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
starlette==0.14.2
text-unidecode==1.3
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
Twisted==21.2.0
typing-extensions==3.10.0.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.13.4
w3lib==1.22.0
whitenoise==5.2.0
zope.interface==5.4.0
